requirements.txt
cryptography
python-gnupg
click
argon2-cffi
keyring
